Game Test Files Steup and Structure
Here are some data structures we will be using for the project. Each is designed to read from a file which contains all information pertaining to their respective class.  The files can be edited through notepad++ or other .csv reader. The main.cpp constructs each class and specific pbjects so you can experiment with adding new things or just figuring out how it all works. Everything needed for the game when it comes to the battle sequences, player phase, enemy phase, etc can be extracted through functions. Need to add more though, so keep an eye out for updates.