Add files via upload
Since we aren't using the older classes anymore I'm gonna have to remake the back end part so I took out all the back end part and am uploading pretty much just the scenes and stuff. To change backgrounds from hospital/mansion/clocktower/church use +, -, /, * on the numpad. the Up arrow was a weak attack and the right arrow is a strong attack.